refactor: abstracted actionbutton into its own file
